+++ module/Application/config/others/these.config.php
@@ -72,6 +72,7 @@ return [
                         //
                         'privileges' => [
                             ThesePrivileges::THESE_SAISIE_CORREC_AUTORISEE_FORCEE,
+                            ThesePrivileges::THESE_CORREC_AUTORISEE_ACCORDER_SURSIS,
                             ThesePrivileges::THESE_SAISIE_DESCRIPTION_VERSION_INITIALE,
                             ThesePrivileges::THESE_SAISIE_DESCRIPTION_VERSION_CORRIGEE,
                             ThesePrivileges::THESE_SAISIE_ATTESTATIONS_VERSION_INITIALE,
@@ -238,6 +239,16 @@ return [
                     ],
                     'privileges' => [
                         ThesePrivileges::THESE_SAISIE_CORREC_AUTORISEE_FORCEE,
+                    ],
+                    'assertion' => 'Assertion\\These',
+                ],
+                [
+                    'controller' => 'Application\Controller\These',
+                    'action' => [
+                        'accorder-sursis-correction',
+                    ],
+                    'privileges' => [
+                        ThesePrivileges::THESE_CORREC_AUTORISEE_ACCORDER_SURSIS,
                     ],
                     'assertion' => 'Assertion\\These',
                 ],
@@ -751,6 +762,7 @@ return [
                                 'these' => '\d+',
                             ],
                             'defaults' => [
+                                /** @see TheseController::accorderSursisCorrectionAction() */
                                 'controller' => 'Application\Controller\These',
                                 'action' => 'validation-these-corrigee',
                             ],
@@ -766,6 +778,19 @@ return [
                             'defaults' => [
                                 'controller' => 'Application\Controller\These',
                                 'action' => 'modifier-correction-autorisee-forcee',
+                            ],
+                        ],
+                    ],
+                    'accorder-sursis-correction' => [
+                        'type' => 'Segment',
+                        'options' => [
+                            'route' => '/accorder-sursis-correction/:these',
+                            'constraints' => [
+                                'these' => '\d+',
+                            ],
+                            'defaults' => [
+                                'controller' => 'Application\Controller\These',
+                                'action' => 'accorder-sursis-correction',
                             ],
                         ],
                     ],