+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Journal des modifications
 - Renommage de l'application en ESUP-SyGAL.
 - Modèle de page de couverture de thèse personnalisable.
 - Substitution possible du favicon.
+- Possibilité d'accorder un sursis pour le dépôt de la verison corrigée de la thèse.
 - [FIX] Bloquage de la signature multiple de l'engagement d'impartilité
 - [FIX] Correction du tri des thèses par titre (plantage) et par date de soutenance (inopérant)
 - Ajout des gardes pour les actions de la page de presoutenance (au cas où l'accés est donnée aux acteurs directs)